requirements.txt
Django==2.2.7
djangorestframework==3.12.2
